start to refactoring the tasks and task factory
